Fixed time charts in timed tabs Interrupting vehicle route if empty and needs despawn from line
Related to #101, except ensuring the travels start from terminal stops
