+++ UI/i18n/en.properties
@@ -464,3 +464,4 @@ TLM_ADDPALETTE_ERROR_INVALIDNAME=<color yellow>Invalid name!</color>
 TLM_TOOLINFO_PREVSTOP=Previous
 TLM_TOOLINFO_NEXTSTOP=Next
 TLM_TOOLINFO_SEGMENTLENGHT=Stops distance
+TLM_ADD_ENTRY=Add entry
