+++ UI/i18n/pt.properties
@@ -465,3 +465,4 @@ TLM_ADDPALETTE_ERROR_INVALIDNAME=<color yellow>Nome inválido</color>
 TLM_TOOLINFO_PREVSTOP=Anterior
 TLM_TOOLINFO_NEXTSTOP=Próxima
 TLM_TOOLINFO_SEGMENTLENGHT=Tamanho do trecho
+TLM_ADD_ENTRY=Adicionar registro
