+++ WorldInfoPanel/Tabs/TLMTicketConfigTab.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 namespace Klyte.TransportLinesManager.UI
 {
 
-    public class TLMTicketConfigTab : TLMTimedConfigTab<TLMTicketConfigTab, TLMTicketPriceEditorLine, TicketPriceEntryXml>
+    public class TLMTicketConfigTab : TLMBaseTimedConfigTab<TLMTicketConfigTab, TLMTicketPriceTimeChart, TLMTicketPriceEditorLine, TicketPriceEntryXml>
     {
         public override string GetTitleLocale() => "K45_TLM_PER_HOUR_TICKET_PRICE_TITLE";
         public override string GetValueColumnLocale() => "K45_TLM_TICKET_PRICE";
